Cleaned up ROC curve plotting scripts
